+++ src/main/java/com/github/unchama/seichiassist/commands/EventCommand.kt
@@ -1,7 +1,7 @@
 package com.github.unchama.seichiassist.commands
 
 import com.github.unchama.contextualexecutor.asNonBlockingTabExecutor
-import com.github.unchama.effect.EmptyEffect
+import com.github.unchama.targetedeffect.EmptyEffect
 import com.github.unchama.seichiassist.commands.contextual.builder.BuilderTemplates.playerCommandBuilder
 import com.github.unchama.seichiassist.listener.new_year_event.NewYearBagListener
 import com.github.unchama.seichiassist.listener.new_year_event.NewYearItemListener
