+++ src/main/java/com/github/unchama/seichiassist/commands/GiganticFeverCommand.kt
@@ -2,7 +2,7 @@ package com.github.unchama.seichiassist.commands
 
 import com.github.unchama.contextualexecutor.asNonBlockingTabExecutor
 import com.github.unchama.contextualexecutor.builder.ContextualExecutorBuilder
-import com.github.unchama.effect.EmptyEffect
+import com.github.unchama.targetedeffect.EmptyEffect
 import com.github.unchama.seichiassist.SeichiAssist
 import com.github.unchama.seichiassist.util.Util
 import kotlinx.coroutines.delay
