+++ src/main/java/com/github/unchama/targetedeffect/TargetedEffect.kt
@@ -1,4 +1,4 @@
-package com.github.unchama.effect
+package com.github.unchama.targetedeffect
 
 import arrow.typeclasses.Monoid
